requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
openai==1.93.0
sentence-transformers==2.6.1
torch==2.7.1+cu118
scikit-learn==1.4.2
numpy==1.26.4
markdown==3.8.2
requests==2.31.0
python-dotenv==1.1.1
httpx==0.28.1
transformers==4.53.0
datasets==2.19.1
python-multipart==0.0.20
jinja2==3.1.6
streamlit==1.46.1